odatests/requirements.txt
pyyaml
pymysql
peewee
flask
ansi2html
